Fixed a bunch of compiler warnings

+++ src/effect/EffectDisintegrate.cpp
@@ -61,7 +61,7 @@ EffectDisintegrate::blit(SDL_Surface *screen, SDL_Surface *surface,
 
     for (int py = 0; py < surface->h; ++py) {
         for (int px = 0; px < surface->w; ++px) {
-            if (Random::aByte(py * surface->w + px) < m_disint) {
+            if (Random::aByte(static_cast<unsigned int>(py * surface->w + px)) < m_disint) {
                 SDL_Color pixel = PixelTool::getColor(surface, px, py);
                 if (pixel.a == 255) {
                     PixelTool::putColor(screen, x + px, y + py, pixel);
